Cancel pending tasks waiting for session loss after lock is released.
Everytime lock is acquired by calling this statement "async with
zk.recipes.Lock(path).acquire():", a task is created waiting for zookeeper
session loss. However that task is still pending after the lock is released. If
a daemon program keeps zookeeper session over a month while it is running,
there exist tons of pending tasks and it leads to resource leakage problem.