Lock: add test code for deadlock incurred by znode creation failure.
While acquiring a lock, create_unique_znode raises an exception but the
ephemeral znode is successfully created in zookeeper.

As one of scenarios, this happens when state changes into SUSPENDED so that the
repair loop aborts the pending creation request. Because the request belongs to
the old connection. However the znode is created successfully before the
request is aborted and newly created znode is not tracked by aiozk. The real
problem is that the untracked znode blocks others not to acquire until it is
deleted by session expiration.

Reviewed-by: Kirill Pinchuk <cybergrind@gmail.com>