Lock: fix deadlock incurred by lock acquisition failure.
While acquiring a lock, session state may change to SUSPENDED. It causes
TimeoutError at .analyze_siblings() method. This makes a znode of
zookeeper be untracked.

Improvements involved in this patch:
Handle TimeoutError raised at .analyze_siblings().
Make lock acquisition not fail if session is suspended but timeout of lock is not reached yet.