barrier: Add test code regarding that a znode becomes untracked
If DoubleBarrier.enter() failed with timeout, a znode which is related to the barrier gets untracked and no objects are responsible for it. So this behavior incurs resource leakage and leads to other tasks to be blocked indefinitely at .leave() method.

The correct behavior should be like this: DoubleBarrier should delete a znode by itself whenever .enter() method failed with any exceptions.