barrier: Delete a znode before returning .enter() if an exception raises
Delete a unique znode created by the barrier if any exceptions raise inside the .enter method. If the znode is leaked, other tasks can proceed with the insufficient number of members. And also the tasks are going to block at .leave() indefinitely because the leaked znode is never going to be deleted.