Replace asynctest with unittest
asynctest seems not maintained for over a year. And when using asynctest dozens
of deprecation warning messages pop out.

Here's a sample of warning messages:
DeprecationWarning: "@coroutine" decorator is deprecated since Python 3.8, use "async def" instead

As of python3.8, AsyncMock is introduced by unittest.mock. So it would be
better to use unittest.mock in lieu of asynctest.